Move application startup to index.js and use /bin/app from template
